+++ app/src/organisms/ProtocolSetup/RunSetupCard/RobotCalibrationStep/index.tsx
@@ -4,14 +4,14 @@ import { Trans, useTranslation } from 'react-i18next'
 
 import { Text, SPACING_3 } from '@opentrons/components'
 
-import type { Dispatch, State } from '../../redux/types'
-import type { ViewableRobot } from '../../redux/discovery/types'
-import * as PipetteOffset from '../../redux/calibration/pipette-offset'
-import * as Pipettes from '../../redux/pipettes'
-import * as PipetteConstants from '../../redux/pipettes/constants'
-import * as TipLength from '../../redux/calibration/tip-length'
+import type { Dispatch, State } from '../../../../redux/types'
+import type { ViewableRobot } from '../../../../redux/discovery/types'
+import * as PipetteOffset from '../../../../redux/calibration/pipette-offset'
+import * as Pipettes from '../../../../redux/pipettes'
+import * as PipetteConstants from '../../../../redux/pipettes/constants'
+import * as TipLength from '../../../../redux/calibration/tip-length'
 import { DeckCalibration } from './DeckCalibration'
-import { formatLastModified } from '../../organisms/CalibrationPanels/utils'
+import { formatLastModified } from '../../../CalibrationPanels/utils'
 
 interface Props {
   robot: ViewableRobot
